Small Fix GP_MapSampler

+++ includes/core_metrics.hpp
@@ -19,7 +19,7 @@ struct Genotype_Metrics
 
   Genotype ref_genotype;
   std::vector <Phenotype_ID> ref_pIDs;
-  std::map <Phenotype_ID, uint16_t> pID_counter;
+  std::map <Phenotype_ID, uint8_t> pID_counter;
 
   Phenotype_ID rare_pID = {0, 0}, loop_pID = {255, 0};
